refactor: Refactor the Launcher Utilities.
