+++ src/ansys/fluent/core/launcher/launcher_utils.py
@@ -1,8 +1,5 @@
 """Provides a module for launching utilities."""
 
-from enum import Enum
-from functools import total_ordering
-import json
 import logging
 import os
 from pathlib import Path
@@ -15,9 +12,10 @@ from typing import Any, Dict, Optional, Union
 
 from beartype import BeartypeConf, beartype
 
-from ansys.fluent.core.exceptions import DisallowedValuesError, InvalidArgument
+from ansys.fluent.core.exceptions import InvalidArgument
 from ansys.fluent.core.fluent_connection import FluentConnection, PortNotProvided
-from ansys.fluent.core.scheduler import build_parallel_options, load_machines
+from ansys.fluent.core.launcher.custom_exceptions import IpPortNotProvided
+from ansys.fluent.core.launcher.pyfluent_enums import FluentMode
 from ansys.fluent.core.session import _parse_server_info_file
 from ansys.fluent.core.session_meshing import Meshing
 from ansys.fluent.core.session_pure_meshing import PureMeshing
@@ -28,57 +26,12 @@ from ansys.fluent.core.utils.fluent_version import FluentVersion
 from ansys.fluent.core.utils.networking import find_remoting_ip
 import ansys.platform.instancemanagement as pypim
 
-_THIS_DIR = os.path.dirname(__file__)
-_OPTIONS_FILE = os.path.join(_THIS_DIR, "fluent_launcher_options.json")
 logger = logging.getLogger("pyfluent.launcher")
-
-
-class InvalidPassword(ValueError):
-    """Raised when password is invalid."""
-
-    def __init__(self):
-        super().__init__("Provide correct 'password'.")
-
-
-class GPUSolverSupportError(ValueError):
-    """Raised when an unsupported Fluent version is specified."""
-
-    def __init__(self):
-        super().__init__("Fluent GPU Solver is only supported for 3D.")
-
-
-class IpPortNotProvided(ValueError):
-    """Raised when ip and port are not specified."""
-
-    def __init__(self):
-        super().__init__("Provide either 'ip' and 'port' or 'server_info_file_name'.")
-
-
-class UnexpectedKeywordArgument(TypeError):
-    """Raised when a valid keyword argument is not specified."""
-
-    pass
-
-
-class DockerContainerLaunchNotSupported(SystemError):
-    """Raised when docker container launch is not supported."""
-
-    def __init__(self):
-        super().__init__("Python Docker SDK is unsupported on this system.")
 
 
 def _is_windows():
     """Check if the current operating system is windows."""
     return platform.system() == "Windows"
-
-
-class LaunchMode(Enum):
-    """An enumeration over supported Fluent launch modes."""
-
-    STANDALONE = 1
-    PIM = 2
-    CONTAINER = 3
-    SLURM = 4
 
 
 def check_docker_support():
@@ -90,37 +43,6 @@ def check_docker_support():
     except docker.errors.DockerException:
         return False
     return True
-
-
-def _get_standalone_launch_fluent_version(
-    product_version: Union[FluentVersion, str, None]
-) -> Optional[FluentVersion]:
-    """Determine the Fluent version during the execution of the ``launch_fluent()``
-    method in standalone mode.
-
-    The search for the version is performed in this order.
-
-    1. The ``product_version`` parameter passed with the ``launch_fluent()`` method.
-    2. The latest Ansys version from ``AWP_ROOTnnn``` environment variables.
-
-    Returns
-    -------
-    FluentVersion, optional
-        Fluent version or ``None``
-    """
-
-    # (DEV) if "PYFLUENT_FLUENT_ROOT" environment variable is defined, we cannot
-    # determine the Fluent version, so returning None.
-    if os.getenv("PYFLUENT_FLUENT_ROOT"):
-        return None
-
-    # Look for Fluent version in the following order:
-    # 1. product_version parameter passed with launch_fluent
-    if product_version:
-        return FluentVersion(product_version)
-
-    # 2. the latest ANSYS version from AWP_ROOT environment variables
-    return FluentVersion.get_latest_installed()
 
 
 def get_fluent_exe_path(**launch_argvals) -> Path:
@@ -162,122 +84,6 @@ def get_fluent_exe_path(**launch_argvals) -> Path:
     return get_exe_path(get_fluent_root(FluentVersion.get_latest_installed()))
 
 
-class FluentMode(Enum):
-    """An enumeration over supported Fluent modes."""
-
-    MESHING_MODE = (Meshing, "meshing")
-    PURE_MESHING_MODE = (PureMeshing, "pure-meshing")
-    SOLVER = (Solver, "solver")
-    SOLVER_ICING = (SolverIcing, "solver-icing")
-
-    @staticmethod
-    def get_mode(mode: str) -> "FluentMode":
-        """Returns the FluentMode based on the provided mode string.
-
-        Parameters
-        ----------
-        mode : str
-            mode
-
-        Returns
-        -------
-        FluentMode
-            Fluent mode
-
-        Raises
-        ------
-        DisallowedValuesError
-            If an unknown mode is passed.
-        """
-        allowed_modes = []
-        for m in FluentMode:
-            allowed_modes.append(m.value[1])
-            if mode == m.value[1]:
-                return m
-        raise DisallowedValuesError("mode", mode, allowed_modes)
-
-    @staticmethod
-    def is_meshing(mode: "FluentMode") -> bool:
-        """Returns whether the current mode is meshing.
-
-        Parameters
-        ----------
-        mode : FluentMode
-            mode
-
-        Returns
-        -------
-        True if mode is FluentMode.MESHING_MODE or FluentMode.PURE_MESHING_MODE else False
-            bool
-        """
-        return mode in [FluentMode.MESHING_MODE, FluentMode.PURE_MESHING_MODE]
-
-
-@total_ordering
-class FluentEnum(Enum):
-    """Provides the base class for Fluent-related enums.
-
-    Accepts lowercase member names as values and supports comparison operators.
-    """
-
-    @classmethod
-    def _missing_(cls, value: str):
-        for member in cls:
-            if str(member) == value:
-                return member
-        raise ValueError(
-            f"The specified value '{value}' is a supported value of {cls.__name__}."
-            f""" The supported values are: '{", '".join(str(member) for member in cls)}'."""
-        )
-
-    def __str__(self):
-        return self.name.lower()
-
-    def __lt__(self, other):
-        if not isinstance(other, type(self)):
-            raise TypeError(
-                f"Cannot compare between {type(self).__name__} and {type(other).__name__}"
-            )
-        if self == other:
-            return False
-        for member in type(self):
-            if self == member:
-                return True
-            if other == member:
-                return False
-
-
-class FluentUI(FluentEnum):
-    """Provides supported user interface mode of Fluent."""
-
-    NO_GUI_OR_GRAPHICS = ("g",)
-    NO_GRAPHICS = ("gr",)
-    NO_GUI = ("gu",)
-    HIDDEN_GUI = ("hidden",)
-    GUI = ("",)
-
-
-class FluentWindowsGraphicsDriver(FluentEnum):
-    """Provides supported graphics driver of Fluent in Windows."""
-
-    NULL = ("null",)
-    MSW = ("msw",)
-    DX11 = ("dx11",)
-    OPENGL2 = ("opengl2",)
-    OPENGL = ("opengl",)
-    AUTO = ("",)
-
-
-class FluentLinuxGraphicsDriver(FluentEnum):
-    """Provides supported graphics driver of Fluent in Linux."""
-
-    NULL = ("null",)
-    X11 = ("x11",)
-    OPENGL2 = ("opengl2",)
-    OPENGL = ("opengl",)
-    AUTO = ("",)
-
-
 def _get_server_info_file_name(use_tmpdir=True):
     server_info_dir = os.getenv("SERVER_INFO_DIR")
     dir_ = (
@@ -316,92 +122,6 @@ def _get_subprocess_kwargs_for_fluent(env: Dict[str, Any], argvals) -> Dict[str,
     return kwargs
 
 
-def _build_fluent_launch_args_string(**kwargs) -> str:
-    """Build Fluent's launch arguments string from keyword arguments.
-
-    Returns
-    -------
-    str
-        Fluent's launch arguments string.
-    """
-    all_options = None
-    with open(_OPTIONS_FILE, encoding="utf-8") as fp:
-        all_options = json.load(fp)
-    launch_args_string = ""
-    for k, v in all_options.items():
-        argval = kwargs.get(k)
-        default = v.get("default")
-        if argval is None and v.get("fluent_required") is True:
-            argval = default
-        if argval is not None:
-            allowed_values = v.get("allowed_values")
-            if allowed_values and argval not in allowed_values:
-                if default is not None:
-                    old_argval = argval
-                    argval = default
-                    logger.warning(
-                        f"Specified value '{old_argval}' for argument '{k}' is not an allowed value ({allowed_values}), default value '{argval}' is going to be used instead."
-                    )
-                else:
-                    logger.warning(
-                        f"{k} = {argval} is discarded as it is not an allowed value. Allowed values: {allowed_values}"
-                    )
-                    continue
-            fluent_map = v.get("fluent_map")
-            if fluent_map:
-                if isinstance(argval, str):
-                    json_key = argval
-                else:
-                    json_key = json.dumps(argval)
-                argval = fluent_map[json_key]
-            launch_args_string += v["fluent_format"].replace("{}", str(argval))
-    addArgs = kwargs.get("additional_arguments")
-    if addArgs and "-t" not in addArgs and "-cnf=" not in addArgs:
-        parallel_options = build_parallel_options(
-            load_machines(ncores=kwargs["processor_count"])
-        )
-        if parallel_options:
-            launch_args_string += " " + parallel_options
-    gpu = kwargs.get("gpu")
-    if gpu is True:
-        launch_args_string += " -gpu"
-    elif isinstance(gpu, list):
-        launch_args_string += f" -gpu={','.join(map(str, gpu))}"
-    ui = kwargs.get("ui")
-    if ui and ui.value[0]:
-        launch_args_string += f" -{ui.value[0]}"
-    graphics_driver = kwargs.get("graphics_driver")
-    if graphics_driver and graphics_driver.value[0]:
-        launch_args_string += f" -driver {graphics_driver.value[0]}"
-    return launch_args_string
-
-
-def _get_mode(mode: Optional[Union[FluentMode, str, None]] = None):
-    """Updates the session information."""
-    if mode is None:
-        mode = FluentMode.SOLVER
-
-    if isinstance(mode, str):
-        mode = FluentMode.get_mode(mode)
-
-    return mode
-
-
-def _raise_non_gui_exception_in_windows(
-    ui: FluentUI, product_version: FluentVersion
-) -> None:
-    """Fluent user interface mode lower than ``FluentUI.HIDDEN_GUI`` is not supported in
-    Windows in Fluent versions lower than 2024 R1."""
-    if (
-        _is_windows()
-        and ui < FluentUI.HIDDEN_GUI
-        and product_version < FluentVersion.v241
-    ):
-        raise InvalidArgument(
-            f"'{ui}' supported in Windows only for Fluent version 24.1 or later."
-        )
-
-
 def _await_fluent_launch(
     server_info_file_name: str, start_timeout: int, sifile_last_mtime: float
 ):
@@ -415,7 +135,7 @@ def _await_fluent_launch(
             raise TimeoutError("The launch process has timed out.")
         time.sleep(1)
         start_timeout -= 1
-        logger.info(f"Waiting for Fluent to launch...")
+        logger.info("Waiting for Fluent to launch...")
         if start_timeout >= 0:
             logger.info(f"...{start_timeout} seconds remaining")
 
@@ -431,7 +151,8 @@ def _get_server_info(
         raise IpPortNotProvided()
     if (ip or port) and server_info_file_name:
         logger.warning(
-            "The ip and port will be extracted from the server-info file and their explicitly specified values will be ignored."
+            "The ip and port will be extracted from the server-info file "
+            "and their explicitly specified values will be ignored."
         )
     else:
         if server_info_file_name:
@@ -443,53 +164,6 @@ def _get_server_info(
         raise PortNotProvided()
 
     return ip, port, password
-
-
-def _get_running_session_mode(
-    fluent_connection: FluentConnection, mode: Optional[FluentMode] = None
-):
-    """Get the mode of the running session if the mode has not been mentioned
-    explicitly."""
-    if mode:
-        session_mode = mode
-    else:
-        try:
-            session_mode = FluentMode.get_mode(
-                "solver"
-                if fluent_connection.scheme_eval.scheme_eval("(cx-solver-mode?)")
-                else "meshing"
-            )
-        except Exception as ex:
-            raise InvalidPassword() from ex
-    return session_mode.value[0]
-
-
-def _generate_launch_string(
-    argvals,
-    mode: FluentMode,
-    additional_arguments: str,
-    server_info_file_name: str,
-):
-    """Generates the launch string to launch fluent."""
-    if _is_windows():
-        exe_path = str(get_fluent_exe_path(**argvals))
-        if " " in exe_path:
-            exe_path = '"' + exe_path + '"'
-    else:
-        exe_path = str(get_fluent_exe_path(**argvals))
-    launch_string = exe_path
-    if mode == FluentMode.SOLVER_ICING:
-        argvals["fluent_icing"] = True
-    launch_string += _build_fluent_launch_args_string(**argvals)
-    if FluentMode.is_meshing(mode):
-        launch_string += " -meshing"
-    if additional_arguments:
-        launch_string += f" {additional_arguments}"
-    if " " in server_info_file_name:
-        server_info_file_name = '"' + server_info_file_name + '"'
-    launch_string += f" -sifile={server_info_file_name}"
-    launch_string += " -nm"
-    return launch_string
 
 
 def _confirm_watchdog_start(start_watchdog, cleanup_on_exit, fluent_connection):
@@ -527,111 +201,6 @@ def _build_journal_argument(
         else:
             fluent_jou_arg += " -topy"
     return fluent_jou_arg
-
-
-# pylint: disable=missing-raises-doc
-class LaunchFluentError(Exception):
-    """Exception class representing launch errors."""
-
-    def __init__(self, launch_string):
-        """__init__ method of LaunchFluentError class."""
-        details = "\n" + "Fluent Launch string: " + launch_string
-        super().__init__(details)
-
-
-def _process_kwargs(kwargs):
-    """Verify whether keyword arguments are valid or not.
-
-    Parameters
-    ----------
-    kwargs: Any
-        Provided keyword arguments.
-
-    Raises
-    ------
-    UnexpectedKeywordArgument
-        If an unexpected keyword argument is provided.
-    """
-    if kwargs:
-        if "meshing_mode" in kwargs:
-            raise UnexpectedKeywordArgument(
-                "Use 'launch_fluent(mode='meshing')' to launch Fluent in meshing mode."
-            )
-        else:
-            raise UnexpectedKeywordArgument(
-                f"launch_fluent() got an unexpected keyword argument {next(iter(kwargs))}"
-            )
-
-
-def _get_fluent_launch_mode(start_container, container_dict, scheduler_options):
-    """Get Fluent launch mode.
-
-    Parameters
-    ----------
-    start_container: bool
-        Specifies whether to launch a Fluent Docker container image.
-    container_dict: dict
-        Dictionary for Fluent Docker container configuration.
-
-    Returns
-    -------
-    fluent_launch_mode: LaunchMode
-        Fluent launch mode.
-    """
-    if pypim.is_configured():
-        fluent_launch_mode = LaunchMode.PIM
-    elif start_container is True or (
-        start_container is None
-        and (container_dict or os.getenv("PYFLUENT_LAUNCH_CONTAINER") == "1")
-    ):
-        if check_docker_support():
-            fluent_launch_mode = LaunchMode.CONTAINER
-        else:
-            raise DockerContainerLaunchNotSupported()
-    elif scheduler_options and scheduler_options["scheduler"] == "slurm":
-        fluent_launch_mode = LaunchMode.SLURM
-    else:
-        fluent_launch_mode = LaunchMode.STANDALONE
-    return fluent_launch_mode
-
-
-def _process_invalid_args(dry_run, fluent_launch_mode, argvals):
-    """Get invalid arguments.
-
-    Parameters
-    ----------
-    dry_run: bool
-        If dry running a container start,
-        ``launch_fluent()`` will return the configured ``container_dict``.
-    fluent_launch_mode: LaunchMode
-        Fluent launch mode.
-    argvals: dict
-        Local arguments.
-    """
-    if dry_run and fluent_launch_mode != LaunchMode.CONTAINER:
-        logger.warning(
-            "'dry_run' argument for 'launch_fluent' currently is only "
-            "supported when starting containers."
-        )
-    if fluent_launch_mode != LaunchMode.STANDALONE:
-        arg_names = [
-            "env",
-            "cwd",
-            "topy",
-            "case_file_name",
-            "lightweight_mode",
-            "journal_file_names",
-            "case_data_file_name",
-        ]
-        invalid_arg_names = list(
-            filter(lambda arg_name: argvals[arg_name] is not None, arg_names)
-        )
-        if len(invalid_arg_names) != 0:
-            invalid_str_names = ", ".join(invalid_arg_names)
-            logger.warning(
-                f"These specified arguments are only supported when starting "
-                f"local standalone Fluent clients: {invalid_str_names}."
-            )
 
 
 def launch_remote_fluent(
