+++ src/ansys/fluent/core/launcher/slurm_launcher.py
@@ -36,15 +36,15 @@ import time
 from typing import Any, Callable, Union
 
 from ansys.fluent.core.exceptions import InvalidArgument
+from ansys.fluent.core.launcher.custom_exceptions import _process_invalid_args
 from ansys.fluent.core.launcher.launcher_utils import (
     _await_fluent_launch,
     _build_journal_argument,
-    _generate_launch_string,
-    _get_mode,
     _get_server_info_file_name,
     _get_subprocess_kwargs_for_fluent,
-    _process_invalid_args,
 )
+from ansys.fluent.core.launcher.process_launch_string import _generate_launch_string
+from ansys.fluent.core.launcher.pyfluent_enums import _get_mode
 from ansys.fluent.core.session_meshing import Meshing
 from ansys.fluent.core.session_pure_meshing import PureMeshing
 from ansys.fluent.core.session_solver import Solver
