+++ tests/test_launcher.py
@@ -8,23 +8,29 @@ import ansys.fluent.core as pyfluent
 from ansys.fluent.core import PyFluentDeprecationWarning  # noqa: F401
 from ansys.fluent.core.exceptions import DisallowedValuesError, InvalidArgument
 from ansys.fluent.core.launcher import launcher_utils
+from ansys.fluent.core.launcher.custom_exceptions import (
+    DockerContainerLaunchNotSupported,
+    GPUSolverSupportError,
+    LaunchFluentError,
+    UnexpectedKeywordArgument,
+    _raise_non_gui_exception_in_windows,
+)
 from ansys.fluent.core.launcher.launcher import create_launcher
 from ansys.fluent.core.launcher.launcher_utils import (
-    DockerContainerLaunchNotSupported,
+    _build_journal_argument,
+    _is_windows,
+    check_docker_support,
+    get_fluent_exe_path,
+)
+from ansys.fluent.core.launcher.process_launch_string import (
+    _build_fluent_launch_args_string,
+)
+from ansys.fluent.core.launcher.pyfluent_enums import (
     FluentLinuxGraphicsDriver,
     FluentMode,
     FluentUI,
     FluentWindowsGraphicsDriver,
-    GPUSolverSupportError,
-    LaunchFluentError,
     LaunchMode,
-    UnexpectedKeywordArgument,
-    _build_fluent_launch_args_string,
-    _build_journal_argument,
-    _is_windows,
-    _raise_non_gui_exception_in_windows,
-    check_docker_support,
-    get_fluent_exe_path,
 )
 from ansys.fluent.core.utils.fluent_version import AnsysVersionNotFound, FluentVersion
 import ansys.platform.instancemanagement as pypim
